feat: 비건 맵 리스트 UseCase적용 및 제보하기, 현재 위치 버튼 추가

+++ data/src/main/java/com/example/data/model/map/VeganMapRestaurantResponse.kt
@@ -11,15 +11,15 @@ data class VeganMapRestaurantResponse(
 
 data class VeganMapRestaurantDto(
     @Json(name = "restaurantId")
-    val restaurantId: Long,
+    val restaurantId: Int,
     @Json(name = "restaurantName")
     val restaurantName: String,
     @Json(name = "restaurantType")
     val restaurantType: String?,
     @Json(name = "distance")
-    val distance: Long,
+    val distance: Double,
     @Json(name = "rate")
-    val rate: Float,
+    val rate: Double,
     @Json(name = "thumbnail")
     val thumbnail: String,
     @Json(name = "latitude")
