+++ domain/src/main/java/com/example/domain/model/map/VeganMapRestaurant.kt
@@ -1,11 +1,11 @@
 package com.example.domain.model.map
 
 data class VeganMapRestaurant(
-    val id: Long,
+    val id: Int,
     val name: String,
     val type: String?,
-    val distance: Long,
-    val rate: Float,
+    val distance: Double,
+    val rate: Double,
     val latitude: String,
     val longitude: String,
     val thumbnail: String
